Completed work with world timer and added cheats for it to CheatManager

+++ Source/AoS_Prototype/Public/AoS_CheatManager.h
@@ -21,7 +21,7 @@ public:
 	UFUNCTION(exec)
 	void CheatSetTimeStamp(FString Day, int32 Hour, int32 Minutes, FString AMPM);
 	UFUNCTION(exec)
-	void CheatIncreaseSunRotationSpeed(int32 Percentage);
+	void CheatIncreaseSunRotationSpeed(float Percentage);
 
 protected:
